Add magic fix for analog O3 runs
This fix adds a nooptimize argument to the boot procedure and also
inserts a stray loop there to wait a few more cycles before
proceeding from boot.

I don't know why this works or how it works, but it works.
We observed this because stepping through the code line-by-line
with GDB worked, but doing a full automatic run didn't.
Since the for loop is only inserted in the boot procedure (which
is called only once in the beginning, it does not affect normal
operation in any way, but it might make boot a little slower.